Refactor

+++ src/main/java/com/github/DKowalski25/_min/service/calendar/CalendarEventServiceImpl.java
@@ -4,6 +4,7 @@ import com.github.DKowalski25._min.dto.calendar.CalendarEventMapper;
 import com.github.DKowalski25._min.dto.calendar.CalendarEventRequestDTO;
 import com.github.DKowalski25._min.dto.calendar.CalendarEventResponseDTO;
 import com.github.DKowalski25._min.dto.calendar.CalendarEventUpdateDTO;
+import com.github.DKowalski25._min.exceptions.AccessDeniedException;
 import com.github.DKowalski25._min.exceptions.EntityNotFoundException;
 import com.github.DKowalski25._min.models.CalendarEvent;
 import com.github.DKowalski25._min.models.User;
@@ -13,8 +14,10 @@ import com.github.DKowalski25._min.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,14 +28,16 @@ public class CalendarEventServiceImpl implements CalendarEventService {
     private final CalendarEventMapper calendarEventMapper;
 
     @Override
-    public List<CalendarEventResponseDTO> getUserEvents(int userId) {
+    @Transactional(readOnly = true)
+    public List<CalendarEventResponseDTO> getUserEvents(UUID userId) {
         return calendarEventRepository.findByUserId(userId).stream()
                 .map(calendarEventMapper::toResponse)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public CalendarEventResponseDTO createEvent(CalendarEventRequestDTO calendarEventRequestDTO, int userId) {
+    @Transactional
+    public CalendarEventResponseDTO createEvent(CalendarEventRequestDTO calendarEventRequestDTO, UUID userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
 
@@ -44,9 +49,14 @@ public class CalendarEventServiceImpl implements CalendarEventService {
     }
 
     @Override
-    public CalendarEventResponseDTO updateEvent(CalendarEventUpdateDTO calendarEventUpdateDTO, int eventId) {
+    @Transactional
+    public CalendarEventResponseDTO updateEvent(CalendarEventUpdateDTO calendarEventUpdateDTO, UUID eventId, UUID userId) {
         CalendarEvent existingCalendarEvent = calendarEventRepository.findById(eventId)
                 .orElseThrow(() -> new EntityNotFoundException("Calendar event not found", eventId));
+
+        if (!existingCalendarEvent.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("You don't have permission to update this event");
+        }
 
         calendarEventMapper.updateEventFromDto(calendarEventUpdateDTO, existingCalendarEvent);
         calendarEventRepository.save(existingCalendarEvent);
@@ -54,7 +64,14 @@ public class CalendarEventServiceImpl implements CalendarEventService {
     }
 
     @Override
-    public void deleteEvent(int eventId) {
+    @Transactional
+    public void deleteEvent(UUID eventId, UUID userId) {
+        CalendarEvent event = calendarEventRepository.findById(eventId)
+                .orElseThrow(() -> new EntityNotFoundException("Event not found"));
+
+        if (!event.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("You don't have permission to update this event");
+        }
         calendarEventRepository.deleteEventById(eventId);
     }
 }
